Update drive-get.md
The line below seems to be in error. It is misleading, customers are questioning the documentation. If the below line is correct, need confirmation from product group if it's possible to provision someone else's OneDrive with delegated permissions and how. Several forums posts and my testing suggest that it is not possible. 

"If a user's OneDrive is not provisioned but the user has a license to use OneDrive, this request will automatically provision the user's drive, when using delegated authentication."